requirements.txt
python-Levenshtein==0.25.1
colorama==0.4.6
requests==2.31.0
packaging==23.2